Minimize code duplication.
