+++ real_intent/events/perplexity/__init__.py
@@ -1,67 +1,11 @@
 """Implementation of event generation using Perplexity."""
 import requests
-from pydantic import ValidationError
-
-from typing import Any, Callable
 import datetime as dt
-import json
 
 from real_intent.events.base import Event, EventsResponse, BaseEventsGenerator
+from real_intent.events.errors import NoValidJSONError, NoEventsFoundError
+from real_intent.events.utils import extract_json_only, retry_generation
 from real_intent.internal_logging import log
-
-
-# ---- Errors ----
-
-class NoValidJSONError(ValueError):
-    """Exception raised when no valid JSON is found in the response."""
-
-    def __init__(self, content: str):
-        super().__init__(content)
-
-
-class NoEventsFoundError(Exception):
-    """Exception raised when no events are found for a zip code."""
-
-    def __init__(self, zip_code: str):
-        super().__init__(f"No events found for zip code {zip_code}")
-
-
-# ---- Helpers ----
-
-def retry_generation(func: Callable):
-    """Retry the generation four times if it fails validation."""
-    MAX_ATTEMPTS: int = 4
-
-    def wrapper(*args, **kwargs):
-        """Run the function, catch error, then retry up to four times."""
-        for attempt in range(1, MAX_ATTEMPTS+1):
-            try:
-                return func(*args, **kwargs)
-            except (ValidationError, KeyError, NoValidJSONError, json.decoder.JSONDecodeError):
-                if attempt < MAX_ATTEMPTS:  # Log warning for first n-1 attempts
-                    log("warn", f"Function {func.__name__} failed validation, attempt {attempt} of {MAX_ATTEMPTS}.")
-                else:  # Log error for the last attempt
-                    log("error", f"Function {func.__name__} failed validation after {MAX_ATTEMPTS} attempts.")
-        
-        # If we've exhausted all attempts, raise the last exception
-        raise
-
-    return wrapper
-
-
-def extract_json_only(response_str: str) -> dict[str, Any]:
-    """
-    Parse a string response and pull out everything between the first { and last }
-    then return it as a dictionary. Allows excess text before and after valid JSON
-    without causing an error.
-    """
-    start_index = response_str.find("{")
-    end_index = response_str.rfind("}")
-
-    if start_index == -1 or end_index == -1:
-        raise NoValidJSONError(response_str)
-
-    return json.loads(response_str[start_index:end_index+1])
 
 
 # ---- Implementation ----
