+++ real_intent/events/utils.py
@@ -1,5 +1,11 @@
 """Utilities for event generation."""
+import json
+from typing import Any, Callable
+from pydantic import ValidationError
+
 from reportlab.lib.pagesizes import letter
+
+from real_intent.events.errors import NoValidJSONError
 from reportlab.lib import colors
 from reportlab.pdfgen import canvas
 from reportlab.platypus import Paragraph
@@ -8,12 +14,49 @@ from io import BytesIO
 
 from real_intent.internal_logging import log
 
+
+def extract_json_only(response_str: str) -> dict[str, Any]:
+    """
+    Parse a string response and pull out everything between the first { and last }
+    then return it as a dictionary. Allows excess text before and after valid JSON
+    without causing an error.
+    """
+    start_index = response_str.find("{")
+    end_index = response_str.rfind("}")
+
+    if start_index == -1 or end_index == -1:
+        raise NoValidJSONError(response_str)
+
+    return json.loads(response_str[start_index:end_index+1])
+
+
+def retry_generation(func: Callable):
+    """Retry the generation four times if it fails validation."""
+    MAX_ATTEMPTS: int = 4
+
+    def wrapper(*args, **kwargs):
+        """Run the function, catch error, then retry up to four times."""
+        for attempt in range(1, MAX_ATTEMPTS+1):
+            try:
+                return func(*args, **kwargs)
+            except (ValidationError, KeyError, NoValidJSONError, json.decoder.JSONDecodeError):
+                if attempt < MAX_ATTEMPTS:  # Log warning for first n-1 attempts
+                    log("warn", f"Function {func.__name__} failed validation, attempt {attempt} of {MAX_ATTEMPTS}.")
+                else:  # Log error for the last attempt
+                    log("error", f"Function {func.__name__} failed validation after {MAX_ATTEMPTS} attempts.")
+        
+        # If we've exhausted all attempts, raise the last exception
+        raise
+
+    return wrapper
+
+
 from real_intent.events.base import EventsResponse
 
 
 # ---- PDF ----
 
-def generate_pdf_buffer(self, events_response: EventsResponse) -> BytesIO:
+def generate_pdf_buffer(events_response: EventsResponse) -> BytesIO:
     """
     Generate a PDF file with the events and summary.
     """
